// code/frontend/frp-fe/src/api/models/ErrorDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Standard error response object
 */
export type ErrorDto = {
    /**
     * Type of the error
     */
    type?: string;
    /**
     * Error message
     */
    message?: string;
    /**
     * Stack trace (only in dev mode)
     */
    stackTrace?: string;
};